fix(spatial-audio): resolve additional UE 5.6 compilation errors
- Make FSpatialSpeakerDSPManager non-copyable/move-only (TUniquePtr array)
- Fix int64 shift undefined behavior in OSC serialization
- Fix int to EDS100MappingArea enum conversion
- Include full SpatialRenderingEngine.h for UHT-generated code

+++ Plugins/RshipSpatialAudio/Source/RshipSpatialAudioRuntime/Private/ExternalProcessor/ExternalProcessorTypes.cpp
@@ -198,7 +198,7 @@ TArray<uint8> FSpatialOSCMessage::Serialize() const
 
 		case ESpatialOSCArgumentType::Int64:
 			// 64-bit big-endian
-			WriteInt32BE(Buffer, static_cast<int32>(Arg.IntValue >> 32));
+			WriteInt32BE(Buffer, static_cast<int32>(static_cast<int64>(Arg.IntValue) >> 32));
 			WriteInt32BE(Buffer, static_cast<int32>(Arg.IntValue & 0xFFFFFFFF));
 			break;
